String Length
Realized I forgot the string length part of the format when writing strings. The other options is that I could use MUTF-8 strings, and end them with a null-terminator instead, deciding if I should do that maybe. I am going to use VarInt/VarLong for numbers anyways, so I think using a length-prefix probably isn't that big of a deal, but we'll see. Adding the string lengths definitely makes it a much bigger file.

https://stackoverflow.com/questions/24614553/why-is-varint-an-efficient-data-representation