FIFO Reviver Issues
The replacer revolves around having FIFO rules to keep track of the keys on the way in to the serialization, and going back the other way would also mean that the values are read the same order again too. This however appears not to be the case. I need to find a new way to properly save the order of things, but backwards to the other, so LIFO?